[Grid] Move source into outgoing/incoming transaction (#115)
This moves source out of the base transaction object and into the incoming/outgoing transaction. We still enforce it for outgoing transaction but not for a incoming transaction. The reason why we need to do this is because an incoming transaction like a funding event will not have a source